Implementando mais testes

+++ src/main/java/br/unitins/topicos1/dto/HardwareDTO.java
@@ -8,5 +8,5 @@ public record HardwareDTO(
         Date lancamento,
         Integer nivel,
         Integer integridade,
-        Integer idFabricante) {
+        Long idFabricante) {
 }
